change folders structure
