--- src/main/java/Main.java
@@ -1,10 +1,0 @@
-import instances.configuration.InstancesConfiguration;
-import white.ra.simple.io.container.DependencyInjectionContainer;
-
-public class Main {
-    public static void main(String[] args) {
-        DependencyInjectionContainer dependencyInjectionContainer = new DependencyInjectionContainer();
-        dependencyInjectionContainer.configureFromJavaInstanceConfiguration(new InstancesConfiguration());
-        dependencyInjectionContainer.initInstances();
-    }
-}
