--- src/main/java/contract/constant/XOCanvasConst.java
@@ -1,8 +1,0 @@
-package contract.constant;
-
-public final class XOCanvasConst {
-    public final static Integer WIDTH = 500;
-    public final static Integer HEIGHT = 500;
-
-    private XOCanvasConst() { }
-}
