--- src/main/java/contract/entity/MapInterface.java
@@ -1,7 +1,0 @@
-package contract.entity;
-
-import contract.enums.PlayerMarkEnum;
-
-public interface MapInterface {
-    PlayerMarkEnum[] getField();
-}
