--- src/main/java/contract/entity/ScoreInterface.java
@@ -1,6 +1,0 @@
-package contract.entity;
-
-public interface ScoreInterface {
-    Integer getXPlayer();
-    Integer getOPlayer();
-}
