--- src/main/java/contract/enums/PlayerMarkEnum.java
@@ -1,6 +1,0 @@
-package contract.enums;
-
-public enum PlayerMarkEnum {
-    X,
-    O;
-}
