--- src/main/java/model/factory/MapFactory.java
@@ -1,9 +1,0 @@
-package model.factory;
-
-import model.entity.Map;
-
-public class MapFactory {
-    public Map create() {
-        return new Map();
-    }
-}
