--- src/main/java/presenter/MapPresenterInterface.java
@@ -1,7 +1,0 @@
-package presenter;
-
-import contract.entity.MapInterface;
-
-public interface MapPresenterInterface {
-    public MapInterface getMap();
-}
