--- src/main/java/presenter/ScorePresenterInterface.java
@@ -1,7 +1,0 @@
-package presenter;
-
-import contract.entity.ScoreInterface;
-
-public interface ScorePresenterInterface {
-    ScoreInterface getScore();
-}
