--- src/main/java/view/contract/ControlPanelInterface.java
@@ -1,7 +1,0 @@
-package view.contract;
-
-import javax.swing.*;
-
-public interface ControlPanelInterface {
-    JButton getRestartButton();
-}
