--- src/main/java/view/contract/ScorePanelInterface.java
@@ -1,5 +1,0 @@
-package view.contract;
-
-public interface ScorePanelInterface {
-    void updateScore();
-}
