--- src/main/java/view/contract/XOCanvasInterface.java
@@ -1,5 +1,0 @@
-package view.contract;
-
-public interface XOCanvasInterface {
-    void repaint();
-}
