--- src/main/java/view/exception/IncorrectMarkValueException.java
@@ -1,5 +1,0 @@
-package view.exception;
-
-public class IncorrectMarkValueException extends RuntimeException {
-
-}
