--- src/main/java/view/panel/GameFieldPanel.java
@@ -1,22 +1,0 @@
-package view.panel;
-
-import view.canvas.XOCanvas;
-
-import javax.swing.*;
-import javax.swing.border.EtchedBorder;
-
-public class GameFieldPanel extends JPanel {
-    private final XOCanvas xoCanvas;
-
-    public GameFieldPanel(XOCanvas xoCanvas) {
-        this.xoCanvas = xoCanvas;
-
-        setBorder(BorderFactory.createEtchedBorder(EtchedBorder.LOWERED));
-
-        add(this.xoCanvas);
-    }
-
-    public XOCanvas getXoCanvas() {
-        return xoCanvas;
-    }
-}
