+++ src/main/java/whitera/xogamenative/contract/constant/XOCanvasConst.java
@@ -1,0 +1,8 @@
+package whitera.xogamenative.contract.constant;
+
+public final class XOCanvasConst {
+    public final static Integer WIDTH = 500;
+    public final static Integer HEIGHT = 500;
+
+    private XOCanvasConst() { }
+}
