+++ src/main/java/whitera/xogamenative/contract/entity/MapInterface.java
@@ -1,0 +1,7 @@
+package whitera.xogamenative.contract.entity;
+
+import whitera.xogamenative.contract.enums.PlayerMarkEnum;
+
+public interface MapInterface {
+    PlayerMarkEnum[] getField();
+}
