+++ src/main/java/whitera/xogamenative/contract/entity/ScoreInterface.java
@@ -1,0 +1,6 @@
+package whitera.xogamenative.contract.entity;
+
+public interface ScoreInterface {
+    Integer getXPlayer();
+    Integer getOPlayer();
+}
