+++ src/main/java/whitera/xogamenative/contract/enums/PlayerMarkEnum.java
@@ -1,0 +1,6 @@
+package whitera.xogamenative.contract.enums;
+
+public enum PlayerMarkEnum {
+    X,
+    O;
+}
