+++ src/main/java/whitera/xogamenative/model/contract/GameInterface.java
@@ -1,0 +1,14 @@
+package whitera.xogamenative.model.contract;
+
+import whitera.xogamenative.contract.entity.MapInterface;
+import whitera.xogamenative.contract.entity.ScoreInterface;
+
+public interface GameInterface {
+    MapInterface getMap();
+
+    ScoreInterface getScore();
+
+    void restart();
+
+    void doTurn(Integer xCoordinate, Integer yCoordinate);
+}
