+++ src/main/java/whitera/xogamenative/model/factory/MapFactory.java
@@ -1,0 +1,9 @@
+package whitera.xogamenative.model.factory;
+
+import whitera.xogamenative.model.entity.Map;
+
+public class MapFactory {
+    public Map create() {
+        return new Map();
+    }
+}
