+++ src/main/java/whitera/xogamenative/presenter/MapPresenterInterface.java
@@ -1,0 +1,7 @@
+package whitera.xogamenative.presenter;
+
+import whitera.xogamenative.contract.entity.MapInterface;
+
+public interface MapPresenterInterface {
+    public MapInterface getMap();
+}
