+++ src/main/java/whitera/xogamenative/presenter/ScorePresenterInterface.java
@@ -1,0 +1,7 @@
+package whitera.xogamenative.presenter;
+
+import whitera.xogamenative.contract.entity.ScoreInterface;
+
+public interface ScorePresenterInterface {
+    ScoreInterface getScore();
+}
