+++ src/main/java/whitera/xogamenative/view/contract/ControlPanelInterface.java
@@ -1,0 +1,7 @@
+package whitera.xogamenative.view.contract;
+
+import javax.swing.*;
+
+public interface ControlPanelInterface {
+    JButton getRestartButton();
+}
