+++ src/main/java/whitera/xogamenative/view/contract/ScorePanelInterface.java
@@ -1,0 +1,5 @@
+package whitera.xogamenative.view.contract;
+
+public interface ScorePanelInterface {
+    void updateScore();
+}
