+++ src/main/java/whitera/xogamenative/view/contract/XOCanvasInterface.java
@@ -1,0 +1,5 @@
+package whitera.xogamenative.view.contract;
+
+public interface XOCanvasInterface {
+    void repaint();
+}
