+++ src/main/java/whitera/xogamenative/view/exception/IncorrectMarkValueException.java
@@ -1,0 +1,5 @@
+package whitera.xogamenative.view.exception;
+
+public class IncorrectMarkValueException extends RuntimeException {
+
+}
